move CacheFile to own submodule
